+++ components/atom/videoPlayer/demo/index.js
@@ -1,33 +1,35 @@
 import {useState} from 'react'
 
-import AtomVideoPlayer from 'components/atom/videoPlayer/src'
+import {H1} from '@s-ui/documentation-library'
+
+import DemoPlayer from './DemoPlayer.js'
 
 export default () => {
   const [selectedFile, setSelectedFile] = useState('')
 
   return (
-    <div style={{maxWidth: '500px'}}>
-      <h1>YouTube</h1>
-      <AtomVideoPlayer src="https://www.youtube.com/embed/1gI_HGDgG7c" />
+    <div className="DemoVideoPlayer">
+      <H1>YouTube</H1>
+      <DemoPlayer src="https://www.youtube.com/embed/1gI_HGDgG7c" />
 
-      <h1>VIMEO</h1>
-      <AtomVideoPlayer src="https://vimeo.com/54289199" />
+      <H1>VIMEO</H1>
+      <DemoPlayer src="https://vimeo.com/54289199" />
 
-      <h1>Adaptative streaming</h1>
-      <AtomVideoPlayer src="https://media-frontend.yams-pro.mpi-internal.com/api/v1/yams-frontend/statics/vo/surf.mp4/hls.m3u8" />
+      <H1>Adaptative streaming</H1>
+      <DemoPlayer src="https://media-frontend.yams-pro.mpi-internal.com/api/v1/yams-frontend/statics/vo/surf.mp4/hls.m3u8" />
 
-      <h1>Native from remote url</h1>
-      <AtomVideoPlayer src="https://cdn.coverr.co/videos/coverr-boat-in-the-sea-5656/1080p.mp4" />
+      <H1>Native from remote url</H1>
+      <DemoPlayer src="https://cdn.coverr.co/videos/coverr-boat-in-the-sea-5656/1080p.mp4" />
 
-      <h1>Native from a blob</h1>
+      <H1>Native from a blob</H1>
       <input
         type="file"
         accept="video/*"
-        onChange={e => {
+        onChange={event => {
           setSelectedFile(event.target.files[0])
         }}
       />
-      <AtomVideoPlayer src={selectedFile} />
+      <DemoPlayer src={selectedFile} />
     </div>
   )
 }