+++ components/atom/videoPlayer/test/index.test.js
@@ -48,7 +48,7 @@ describe('AtomVideoPlayer', () => {
   })
 
   describe('Error cases', () => {
-    it('should display "Not supported media type" when a unknown video provier is received', () => {
+    it('should display "Not supported media type" when a unknown video provier is received', async () => {
       // Given
       const props = {
         src: 'https://random.and.invalid.video.com/hello-world-video'
@@ -58,12 +58,12 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      component.getByText('Not supported media type')
+      await component.findByText('Not supported media type')
     })
   })
 
   describe('YouTube Videos', () => {
-    it('should embed the youtube video player if src is a youtube video', () => {
+    it('should embed the youtube video player if src is a youtube video', async () => {
       // Given
       const props = {
         src: 'https://www.youtube.com/embed/1gI_HGDgG7c'
@@ -73,10 +73,10 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
     })
 
-    it('should convert standard youtube urls to embedable urls', () => {
+    it('should convert standard youtube urls to embedable urls', async () => {
       // Given
       const props = {
         src: 'https://www.youtube.com/watch?v=1gI_HGDgG7c'
@@ -86,14 +86,14 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include(
         'https://www.youtube.com/embed/1gI_HGDgG7c'
       )
     })
 
-    it('should convert shared videos urls to embedable urls', () => {
+    it('should convert shared videos urls to embedable urls', async () => {
       // Given
       const props = {
         src: 'https://youtu.be/1gI_HGDgG7c'
@@ -103,14 +103,14 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include(
         'https://www.youtube.com/embed/1gI_HGDgG7c'
       )
     })
 
-    it('should avoid displaying controls if controls prop is set to false', () => {
+    it('should avoid displaying controls if controls prop is set to false', async () => {
       // Given
       const props = {
         controls: false,
@@ -121,12 +121,12 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include('controls=0')
     })
 
-    it('should autoplay the video if the autoPlay prop is set to true', () => {
+    it('should autoplay the video if the autoPlay prop is set to true', async () => {
       // Given
       const props = {
         autoPlay: true,
@@ -137,12 +137,12 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include('autoplay=1')
     })
 
-    it('should start the video at the expected time if offset param is passed', () => {
+    it('should start the video at the expected time if offset param is passed', async () => {
       // Given
       const props = {
         timeOffset: 10,
@@ -153,12 +153,12 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include('start=10')
     })
 
-    it('should end the video at the expected time if limit param is passed', () => {
+    it('should end the video at the expected time if limit param is passed', async () => {
       // Given
       const props = {
         timeLimit: 15,
@@ -169,14 +169,30 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(YOUTUBE_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include('end=15')
+    })
+
+    it('should mute the video when the muted prop is set to true', async () => {
+      // Given
+      const props = {
+        muted: true,
+        src: 'https://www.youtube.com/embed/1gI_HGDgG7c'
+      }
+
+      // When
+      const component = setup(props)
+
+      // Then
+      const iframeNode = await component.findByTitle(YOUTUBE_DEFAULT_TITLE)
+      // check that the iframe src is the embedable url
+      expect(iframeNode.src).to.include('mute=1')
     })
   })
 
   describe('VIMEO Videos', () => {
-    it('should embed the vimeo video player if src is a vimeo video', () => {
+    it('should embed the vimeo video player if src is a vimeo video', async () => {
       // Given
       const props = {
         src: 'https://vimeo.com/54289199'
@@ -186,10 +202,10 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      component.getByTitle(VIMEO_DEFAULT_TITLE)
+      await component.findByTitle(VIMEO_DEFAULT_TITLE)
     })
 
-    it('should convert standard vimeo url to an embeddable url', () => {
+    it('should convert standard vimeo url to an embeddable url', async () => {
       // Given
       const props = {
         src: 'https://vimeo.com/54289199'
@@ -199,14 +215,14 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(VIMEO_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(VIMEO_DEFAULT_TITLE)
       // check that the iframe src is the embedable url
       expect(iframeNode.src).to.include(
         'https://player.vimeo.com/video/54289199'
       )
     })
 
-    it('should avoid displaying controls if controls prop is set to false', () => {
+    it('should avoid displaying controls if controls prop is set to false', async () => {
       // Given
       const props = {
         controls: false,
@@ -217,11 +233,11 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(VIMEO_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(VIMEO_DEFAULT_TITLE)
       expect(iframeNode.src).to.include('controls=0')
     })
 
-    it('should autoplay the video if the prop autoPlay is set to true', () => {
+    it('should autoplay the video if the prop autoPlay is set to true', async () => {
       // Given
       const props = {
         autoPlay: true,
@@ -232,11 +248,11 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(VIMEO_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(VIMEO_DEFAULT_TITLE)
       expect(iframeNode.src).to.include('autoplay=1')
     })
 
-    it('should start the video at the expected time if offset param is passed', () => {
+    it('should start the video at the expected time if offset param is passed', async () => {
       // Given
       const props = {
         timeOffset: '25',
@@ -247,13 +263,29 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      const iframeNode = component.getByTitle(VIMEO_DEFAULT_TITLE)
+      const iframeNode = await component.findByTitle(VIMEO_DEFAULT_TITLE)
       expect(iframeNode.src).to.include('#t=25')
+    })
+
+    it('should mute the video when the muted prop is set to true', async () => {
+      // Given
+      const props = {
+        muted: true,
+        src: 'https://vimeo.com/54289199'
+      }
+
+      // When
+      const component = setup(props)
+
+      // Then
+      const iframeNode = await component.findByTitle(VIMEO_DEFAULT_TITLE)
+      // check that the iframe src is the embedable url
+      expect(iframeNode.src).to.include('muted=1')
     })
   })
 
   describe('Adaptative streaming player', () => {
-    it('should open adaptative streaming videos by using hls.js', () => {
+    it('should open adaptative streaming videos by using hls.js', async () => {
       // Given
       const props = {
         src: 'https://media-frontend.yams-pro.mpi-internal.com/api/v1/yams-frontend/statics/vo/surf.mp4/hls.m3u8'
@@ -263,10 +295,10 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      component.getByTitle(HLS_DEFAULT_TITLE)
+      await component.findByTitle(HLS_DEFAULT_TITLE)
     })
 
-    it('should display controls by default', () => {
+    it('should display controls by default', async () => {
       // Given
       const props = {
         src: 'https://media-frontend.yams-pro.mpi-internal.com/api/v1/yams-frontend/statics/vo/surf.mp4/hls.m3u8'
@@ -276,10 +308,11 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      expect(component.getByTitle(HLS_DEFAULT_TITLE).controls).to.eql(true)
+      const player = await component.findByTitle(HLS_DEFAULT_TITLE)
+      expect(player.controls).to.eql(true)
     })
 
-    it('should avoid displaying controls if controls prop is set to false', () => {
+    it('should avoid displaying controls if controls prop is set to false', async () => {
       // Given
       const props = {
         controls: false,
@@ -290,12 +323,28 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      expect(component.getByTitle(HLS_DEFAULT_TITLE).controls).to.eql(false)
+      const player = await component.findByTitle(HLS_DEFAULT_TITLE)
+      expect(player.controls).to.eql(false)
+    })
+
+    it('should mute the video when muted prop is set to true', async () => {
+      // Given
+      const props = {
+        muted: true,
+        src: 'https://media-frontend.yams-pro.mpi-internal.com/api/v1/yams-frontend/statics/vo/surf.mp4/hls.m3u8'
+      }
+
+      // When
+      const component = setup(props)
+
+      // Then
+      const player = await component.findByTitle(HLS_DEFAULT_TITLE)
+      expect(player.muted).to.eql(true)
     })
   })
 
   describe('Native video player', () => {
-    it('should try to play the video using the native player if it is a remote file with a known extension', () => {
+    it('should try to play the video using the native player if it is a remote file with a known extension', async () => {
       // Given
       const props = {
         src: 'https://cdn.coverr.co/videos/coverr-boat-in-the-sea-5656/1080p.mp4'
@@ -305,7 +354,7 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      component.getByTitle(NATIVE_DEFAULT_TITLE)
+      await component.findByTitle(NATIVE_DEFAULT_TITLE)
     })
 
     it('should try to play the video using the native player when receiving a blob object as src', async () => {
@@ -322,7 +371,7 @@ describe('AtomVideoPlayer', () => {
       expect(src).to.include('data:video/mp4;base64,ZmFrZVZpZGVv')
     })
 
-    it('should display controls by default', () => {
+    it('should display controls by default', async () => {
       // Given
       const props = {
         src: 'https://cdn.coverr.co/videos/coverr-boat-in-the-sea-5656/1080p.mp4'
@@ -332,10 +381,11 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      expect(component.getByTitle(NATIVE_DEFAULT_TITLE).controls).to.eql(true)
+      const player = await component.findByTitle(NATIVE_DEFAULT_TITLE)
+      expect(player.controls).to.eql(true)
     })
 
-    it('should avoid displaying controls if controls prop is set to false', () => {
+    it('should avoid displaying controls if controls prop is set to false', async () => {
       // Given
       const props = {
         controls: false,
@@ -346,10 +396,11 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      expect(component.getByTitle(NATIVE_DEFAULT_TITLE).controls).to.eql(false)
+      const player = await component.findByTitle(NATIVE_DEFAULT_TITLE)
+      expect(player.controls).to.eql(false)
     })
 
-    it('should autoplay the video if autoPlay prop is set to true', () => {
+    it('should autoplay the video if autoPlay prop is set to true', async () => {
       // Given
       const props = {
         autoPlay: true,
@@ -360,7 +411,8 @@ describe('AtomVideoPlayer', () => {
       const component = setup(props)
 
       // Then
-      expect(component.getByTitle(NATIVE_DEFAULT_TITLE).autoplay).to.eql(true)
+      const player = await component.findByTitle(NATIVE_DEFAULT_TITLE)
+      expect(player.autoplay).to.eql(true)
     })
 
     it('should start the video at the expected time if offset param is passed', async () => {
@@ -391,6 +443,21 @@ describe('AtomVideoPlayer', () => {
       // Then
       const {src} = await component.findByTestId('videosrc')
       expect(src).to.include('#t=0,20')
+    })
+
+    it('should mute the video when muted prop is set to true', async () => {
+      // Given
+      const props = {
+        muted: true,
+        src: 'https://cdn.coverr.co/videos/coverr-boat-in-the-sea-5656/1080p.mp4'
+      }
+
+      // When
+      const component = setup(props)
+
+      // Then
+      const player = await component.findByTitle(NATIVE_DEFAULT_TITLE)
+      expect(player.muted).to.eql(true)
     })
   })
 })